refactor: useMap custom hook 추가
